# ecommerce/docs/swagger.yaml
definitions:
...
  controllers.saleOrderLineInput:
    properties:
      jumlah_barang:
        type: integer
      product_id:
        type: integer
      sale_order_id:
        type: integer
    type: object
  controllers.uomInput:
    properties:
      nama_unit:
        type: string
    type: object
...
  models.SaleOrderLine:
    properties:
      created_at:
        type: string
      id:
        type: integer
      jumlah_barang:
        type: integer
      jumlah_harga:
        type: integer
      product_id:
        description: FK
        type: integer
      sale_order_id:
        type: integer
      updated_at:
        type: string
    type: object
  models.UnitOfMeasure:
    properties:
      created_at:
        type: string
      id:
        type: integer
      nama_unit:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /product/{id}/sale-order-line:
    get:
      description: Get all SaleOrderLine by Product id.
      parameters:
      - description: Product id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SaleOrderLine'
            type: array
      summary: Get a list of SaleOrderLines.
      tags:
      - Product
...
  /sale-order-line:
    get:
      description: Get a list of SaleOrderLine.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SaleOrderLine'
            type: array
      summary: Get all Sale order line.
      tags:
      - Sale Order Line
    post:
      description: Creating a new SaleOrderLine.
      parameters:
      - description: the body to create a new SaleOrderLine
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.saleOrderLineInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SaleOrderLine'
      security:
      - BearerToken: []
      summary: Create New Review SaleOrderLine.
      tags:
      - Sale Order Line
  /sale-order-line/{id}:
    delete:
      description: Delete a SaleOrderLine by id.
      parameters:
      - description: SaleOrderLine id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one SaleOrderLine.
      tags:
      - Sale Order Line
    put:
      description: Update SaleOrderLine by id.
      parameters:
      - description: SaleOrderLine id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update SaleOrderLine
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.saleOrderLineInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SaleOrderLine'
      security:
      - BearerToken: []
      summary: Update SaleOrderLines.
      tags:
      - Sale Order Line
...
  /sale-order/{id}/sale-order-line:
    get:
      description: Get all SaleOrderLine by SaleOrder id
      parameters:
      - description: SaleOrder id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SaleOrderLine'
            type: array
      summary: Get a list of SaleOrderLines.
      tags:
      - Sale Order
  /uom:
    get:
      description: Get a list of Unit of measures.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UnitOfMeasure'
            type: array
      summary: Get all Unit of measures.
      tags:
      - UoM
    post:
      description: Creating a new Unit of measures.
      parameters:
      - description: the body to create a new UnitOfMeasure
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.uomInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UnitOfMeasure'
      security:
      - BearerToken: []
      summary: Create New Unit of measures.
      tags:
      - UoM
  /uom/{id}:
    delete:
      description: Delete a UoM by id.
      parameters:
      - description: uom id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one Unit of Measures.
      tags:
      - UoM
    put:
      description: Update Unit of measures by id.
      parameters:
      - description: UnitOfMeasure id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update Unit of Measure
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.uomInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UnitOfMeasure'
      security:
      - BearerToken: []
      summary: Update Unit of measures.
      tags:
      - UoM
...
swagger: "2.0"